Query a specific object by setting the id with the Query.id() method

+++ myslicelib/query/__init__.py
@@ -24,6 +24,8 @@ def q(entity: Entity):
 
 class Query(object):
 
+    _id = None
+
     # def __new__(cls, *args, **kwargs):
         # if cls is Query:
         #     try :
@@ -50,92 +52,27 @@ class Query(object):
         self.api = getattr(Api(s.endpoints, s.credential), self.entity.__name__.lower())()
 
 
-        # self.api = Api(s.endpoints, s.credential)
-        # self.api = self.api.resource()
-
-        #for _e,_k in _entities:
-        # res = self.api.version()
-        #
-        # for k in res['ams']:
-        #     print("{} : ".format(k))
-        #
-        # m, c = entity.split('.')
-        # model = 'myslicelib.model.%s' % m
-        # try:
-        #     module = __import__(model, fromlist=[c])
-        # except ImportError:
-        #     logging.error("Model %s not found" % m)
-        #     exit(1)
-        #
-        # try:
-        #     self.cls = getattr(module, c)
-        # except AttributeError as e:
-        #     logging.error(e)
-        #     logging.error("%s entity not found" % (entity.__name__))
-        #     exit(1)
-        #
-        #
-        # self.api = Api(
-        #             'https://www.planet-lab.eu/PLCAPI/',
-        #             'password',
-        #             'myops@planet-lab.eu',
-        #             '1:l-a)be'
-        # )
-        #
-        # self.e = entity
-        # self.f = {}
-        # self.r = None
-
-    # def filter(self, key, value):
-    #     self.f[key] = value
-    #     return self
-    #
-    # def order(self, order):
-    #     return self.sort(order)
-    #
-    # def sort(self, sort):
-    #     return self.filter('-SORT', sort)
-    #
-    # def offset(self, offset):
-    #     if offset > 0:
-    #         return self.filter('-OFFSET', offset)
-    #     return self
-    #
-    # def limit(self, limit):
-    #     return self.filter('-LIMIT', limit)
-    #
-    # def gt(self, key, value):
-    #     return self.filter('>' + key, value)
-    #
-    # def gte(self, key, value):
-    #     return self.filter(']' + key, value)
-    #
-    # def lt(self, key, value):
-    #     return self.filter('<' + key, value)
-    #
-    # def lte(self, key, value):
-    #     return self.filter('[' + key, value)
-    #
-    # def find(self):
-    #     return self.execute()
-
     def id(self, id):
         self._id = id
         return self
 
-    def get(self, id=None):
-        res = self.api.get(id)
+    def get(self):
+        res = self.api.get(self._id)
         for el in res:
             for prop in el:
                 setattr(self.entity, prop, el[prop])
         return self.entity
 
-    def update(self, id, params):
+    def update(self, params):
+        if not self._id:
+            raise Exception("No element specified")
         res = self.api.update(id, params)
         import pprint
         pprint.pprint(res)
 
-    def delete(self, id):
+    def delete(self):
+        if not self._id:
+            raise Exception("No element specified")
         res = self.api.delete(id)
         import pprint
         pprint.pprint(res)
